+++ customer.cpp
@@ -14,7 +14,7 @@ void displayCustomerMenu() {
 
 // Function to handle viewing account details
 void handleViewAccountDetails(BST_Tree& t) {
-    int accountno;
+    long accountno;
     cout << "Please enter account number: ";
     cin >> accountno;
 
@@ -37,7 +37,7 @@ void handleViewAccountDetails(BST_Tree& t) {
 
 // Function to handle viewing transaction history
 void handleViewTransactionHistory() {
-    int accountno;
+    long accountno;
     cout << "Please enter account number: ";
     cin >> accountno;
