+++ staff.cpp
@@ -16,7 +16,7 @@ void displayStaffMenu() {
 
 // Function to handle viewing transaction history
 void handleViewTransactionHistory(BST_Tree& t) {
-    int accountno;
+    long accountno;
     cout << "Please enter account number: ";
     cin >> accountno;
 
@@ -54,7 +54,7 @@ void handleTransfer(BST_Tree& t) {
 
 // Function to handle withdrawing money
 void handleWithdraw(BST_Tree& t) {
-    int accountNo, amount;
+    long accountNo, amount;
 
     cout << "Please enter account number: ";
     cin >> accountNo;
@@ -66,7 +66,7 @@ void handleWithdraw(BST_Tree& t) {
 
 // Function to handle depositing money
 void handleDeposit(BST_Tree& t) {
-    int accountNo, amount;
+    long accountNo, amount;
 
     cout << "Please enter account number: ";
     cin >> accountNo;
